% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_shapes.R
\name{p3d_cube}
\alias{p3d_cube}
\title{Generate random points on the surface of a cube}
\usage{
p3d_cube(n, lims = c(0, 1))
}
\arguments{
\item{n}{Number of points to generate.}

\item{lims}{Vector of min and max ordinates. The default (\code{c(0,1)})
corresponds to a unit cube at the origin.}
}
\value{
A named list with elements:
  \code{coords}: A three column matrix of point coordinates;
  \code{sides}: A vector of integers (values 0-5) identifying the side on
  which each point lies.
}
\description{
Generate random points on the surface of a cube
}
\examples{
xcube <- p3d_cube(n = 1000)
head(xcube$coords)

\dontrun{
# If you have installed package 'threejs' you can view
# and rotate the cube points in a browser window
threejs::scatterplot3js(xcube$coords, size = 0.2)
}

}
