% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_shapes.R
\name{p3d_cylinder}
\alias{p3d_cylinder}
\title{Generate random points on the surface of a cylinder}
\usage{
p3d_cylinder(n, radius, height, pos = c(0, 0, 0), align = "z")
}
\arguments{
\item{n}{Number of points to generate.}

\item{radius}{Cylinder radius (must be positive). Default is 1.}

\item{height}{Cylinder height (must be positive). Default is 1.}

\item{pos}{Either a numeric vector of three values giving the position of the
cylinder centroid on each axis, or a single value which will be used for
all axes. Default is the origin: \code{c(0,0,0)}.}

\item{align}{The axis along which the cylinder should be aligned, specified
as either character (one of (\code{'x', 'y', 'z'}), case ignored) or an
integer (one of \code{1, 2, 3}).}
}
\value{
A named list with elements:
  \describe{
  \item{coords}{A three column matrix of point coordinates.}
  \item{npts}{Number of points.}
  \item{radius}{Cylinder radius.}
  \item{height}{Cylinder height.}
  }
}
\description{
Presently this function only creates a cylinder aligned with one of the axes.
The next step is to allow the orientation of the cylinder to be varied more
flexibly.
}
\examples{
# 4000 points on a cylinder with radius=10 and height=20
# aligned with the X axis.
xcyl <- p3d_cylinder(n = 4000, radius = 10, height = 20, align = "x")
head(xcyl$coords)

\dontrun{
# If you have installed package 'threejs' you can view
# and rotate the sphere points in a browser window
threejs::scatterplot3js(xcyl$coords, size = 0.2)
}

}
