% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_shapes.R
\name{p3d_ellipsoid}
\alias{p3d_ellipsoid}
\title{Generate random points on the surface of an ellipsoid}
\usage{
p3d_ellipsoid(n, radii = c(1, 1, 1), pos = c(0, 0, 0))
}
\arguments{
\item{n}{Number of points to generate.}

\item{radii}{Radii along each of the X, Y and Z axes. Default is c(1,1,1)
which gives a unit sphere. Values must be positive.}

\item{pos}{Either a numeric vector of three values giving the position on
each axis, or a single value which will be used for all axes. Default is
the origin: \code{c(0,0,0)}.}
}
\value{
A named list with elements:
  \describe{
  \item{coords}{A three column matrix of point coordinates.}
  \item{npts}{Number of points.}
  \item{radii}{Vector of length three giving the radius along each axis.}
  }
}
\description{
Generate random points on the surface of an ellipsoid
}
\examples{
# 1000 points on an ellipsoid at the origin
xellipsoid <- p3d_ellipsoid(n = 1000, radii = c(10, 4, 2))
head(xellipsoid$coords)

# 4000 points on a larger ellipsoid at x=10, y=-20, z=0
xellipsoid <- p3d_ellipsoid(n = 4000, radii = c(20, 10, 5), pos = c(10, -20, 0))

\dontrun{
# If you have installed package 'threejs' you can view
# and rotate the ellipsoid in a browser window
threejs::scatterplot3js(xellipsoid$coords, size = 0.2)
}

}
