% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{p3d_rotate}
\alias{p3d_rotate}
\title{Rotate a point cloud}
\usage{
p3d_rotate(x, theta = c(0, 0, 0), origin = "centroid")
}
\arguments{
\item{x}{Either a list containing a 3-column matrix of point coordinates
as a named element \code{'coords'}; or a matrix directly.}

\item{theta}{A numeric vector giving the angle of rotation on the
X, Y and Z axes respectively. If \code{length(theta) < 3} the implied
missing values are set to zero; e.g. \code{theta = pi/4} would result
in rotation around the X-axis but none around the Y or Z axes.}

\item{origin}{The point about which to rotate the point cloud, specified as
either a character string or a 3-element numeric vector of point
coordinates. If a character string: \code{'centroid'} (default) means
rotate around the centroid of the point cloud; \code{'origin'} means rotate
around the origin.}
}
\value{
Rotated point coordinates as a three-column matrix.
}
\description{
Rotate a point cloud
}
