% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{p3d_rz}
\alias{p3d_rz}
\title{Get matrix for rotation around the Z axis}
\usage{
p3d_rz(theta)
}
\arguments{
\item{theta}{Angle of rotation in radians}
}
\value{
3x3 matrix rotation matrix
}
\description{
Get matrix for rotation around the Z axis
}
