% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_shapes.R
\name{p3d_cube}
\alias{p3d_cube}
\title{Generate random points on the surface of a cube}
\usage{
p3d_cube(n, sidelen = 1, pos = c(0, 0, 0))
}
\arguments{
\item{n}{Number of points to generate.}

\item{sidelen}{Cube side length. Default is 1. Negative values are permitted.}

\item{pos}{Either a numeric vector of three values giving the position on
each axis, or a single value which will be used for all axes. Default is
the origin: \code{c(0,0,0)}. If \code{sidelen} is a positive value,
\code{pos} will be the minimum corner position on each axis.}
}
\value{
A named list with elements:
  \describe{
  \item{coords}{A three column matrix of point coordinates.}
  \item{npts}{Number of points.}
  \item{sidelen}{Side length.}
  \item{pos}{Vector of length three giving the position on each axis.}
  \item{sides}{A vector of integers (values 0-5) identifying the side on
    which each point lies.}
  }
}
\description{
Generate random points on the surface of a cube
}
\examples{
# 1000 points on a unit cube at the origin
xcube <- p3d_cube(n = 1000)
head(xcube$coords)

# 4000 points on a larger cube at x=2, y=-5, z=0
xcube <- p3d_cube(n = 4000, sidelen = 5, pos = c(2, -5, 0))

\dontrun{
# If you have installed package 'threejs' you can view
# and rotate the cube points in a browser window
threejs::scatterplot3js(xcube$coords, size = 0.2)
}

}
